Swizzle ActionCable internal disconnect handling in dev/test
This is triggered by calling `ActionCable.server.remote_connections.where(foo: true).disconnect` on the server, which is useful in development & tests to ensure logic works after reconnection.

That is slow because it calls `Connection.close()` which closes the websocket and relies on `ConnectionMonitor` polling to fire once two heartbeats have been missed after ~6-8 seconds. Instead we want to call `Connection.reopen()` which closes the websocket, waits 500ms and opens it again. Same behaviour from the application's point of view, but much quicker from the human's point of view.